+++ packages/server/src/Hocuspocus.ts
@@ -1,4 +1,4 @@
-import type { IncomingMessage } from "http";
+import type { IncomingMessage } from "node:http";
 import { ResetConnection, awarenessStatesToArray } from "@hocuspocus/common";
 import { v4 as uuid } from "uuid";
 import type WebSocket from "ws";
@@ -117,6 +117,7 @@ export class Hocuspocus {
 			afterStoreDocument: this.configuration.afterStoreDocument,
 			onAwarenessUpdate: this.configuration.onAwarenessUpdate,
 			onRequest: this.configuration.onRequest,
+			beforeUnloadDocument: this.configuration.beforeUnloadDocument,
 			afterUnloadDocument: this.configuration.afterUnloadDocument,
 			onDisconnect: this.configuration.onDisconnect,
 			onDestroy: this.configuration.onDestroy,
@@ -491,7 +492,15 @@ export class Hocuspocus {
 		const documentName = document.name;
 		if (!this.documents.has(documentName)) return;
 
-		await this.hooks("beforeUnloadDocument", { instance: this, documentName });
+		try {
+			await this.hooks("beforeUnloadDocument", {
+				instance: this,
+				documentName,
+				document,
+			});
+		} catch (e) {
+			return;
+		}
 
 		if (document.getConnectionsCount() > 0) {
 			return;