+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.2.0](https://github.com/ueberdosis/hocuspocus/compare/v3.1.12...v3.2.0) (2025-07-12)
+
+**Note:** Version bump only for package @hocuspocus/cli
+
+
+
+
+
+## [3.1.12](https://github.com/ueberdosis/hocuspocus/compare/v3.1.11...v3.1.12) (2025-07-09)
+
+**Note:** Version bump only for package @hocuspocus/cli
+
+
+
+
+
+## [3.1.11](https://github.com/ueberdosis/hocuspocus/compare/v3.1.9...v3.1.11) (2025-07-08)
+
+**Note:** Version bump only for package @hocuspocus/cli
+
+
+
+
+
+## [3.1.10](https://github.com/ueberdosis/hocuspocus/compare/v3.1.9...v3.1.10) (2025-07-08)
+
+**Note:** Version bump only for package @hocuspocus/cli
+
+
+
+
+
+## [3.1.9](https://github.com/ueberdosis/hocuspocus/compare/v3.1.8...v3.1.9) (2025-07-01)
+
+**Note:** Version bump only for package @hocuspocus/cli
+
+
+
+
+
+## [3.1.8](https://github.com/ueberdosis/hocuspocus/compare/v3.1.7...v3.1.8) (2025-07-01)
+
+**Note:** Version bump only for package @hocuspocus/cli
+
+
+
+
+
 ## [3.1.7](https://github.com/ueberdosis/hocuspocus/compare/v3.1.6...v3.1.7) (2025-06-22)
 
 **Note:** Version bump only for package @hocuspocus/cli